+++ src/main/java/com/example/kostplan/util/HealthUtil.java
@@ -15,7 +15,7 @@ public class HealthUtil {
 	 * Calculates the BMR (Basal Metabolic Rate) using the Harris-Benedict equation.
 	 * @return the BMR in kilocalories.
 	 */
-	public static int calculateBMR(boolean male, int weight, int height, LocalDate dob) {
+	public static double calculateBMR(boolean male, int weight, int height, LocalDate dob) {
 		double weightFactor = male ? 13.752 : 9.563;
 		double heightFactor = male ? 5 : 1.85;
 		double ageFactor = male ? 6.755 : 4.676;
@@ -28,7 +28,7 @@ public class HealthUtil {
 			- (ageFactor * ageInYears)
 			+ constantTerm;
 
-		return (int)bmr;
+		return bmr;
 	}
 
 	/**
@@ -38,7 +38,7 @@ public class HealthUtil {
 	 * @param calorieGoal The person's calorie goal.
 	 * @param recipe The recipe to scale up or down.
 	 */
-	public static void scaleRecipe(int calorieGoal, Recipe recipe, List<Recipe> scheduledMeals)
+	public static void scaleRecipe(double calorieGoal, Recipe recipe, List<Recipe> scheduledMeals)
 		throws IllegalArgumentException
 	{
 		if (recipe == null || scheduledMeals == null) {
@@ -50,7 +50,7 @@ public class HealthUtil {
 			.map(Recipe::sumCalories)
 			.reduce(Double::sum)
 			.orElse(0.0);
-		double scaleFactor = (double) calorieGoal / totalCalories;
+		double scaleFactor = calorieGoal / totalCalories;
 
 		for (Ingredient ingredient : recipe.getIngredients()) {
 			double newQuantity = ingredient.getQuantity() * scaleFactor;
